requirements.txt
bottle==0.12.13
cGraph==0.1
git+git://github.com/pygraphviz/pygraphviz.git --install-option="--include-path=/usr/include/graphviz" --install-option="--library-path=/app/.heroku/python/lib"
setuptools==39.0.1
requests==2.20.1
transitions==0.6.9
pillow==5.3.0
